added order success page.

+++ src/main/java/com/eventpool/order/service/OrderServiceImpl.java
@@ -160,6 +160,7 @@ public class OrderServiceImpl implements OrderService {
 		TicketInventoryDetails updateTicketInventory = updateTicketInventory(suborderDTO);
 
 		if (updateTicketInventory.isInvUpdated()) {
+			if(suborderDTO.getTicketRegisterId()!=null)
 			ticketRegisterRepository.delete(suborderDTO.getTicketRegisterId());
 		}
